--- client_code/Tabulator/__StartupForm/__init__.py
@@ -1,26 +1,0 @@
-# SPDX-License-Identifier: MIT
-# Copyright (c) 2022 Stu Cork
-
-from anvil import *
-
-from ._anvil_designer import __StartupFormTemplate
-
-
-class __StartupForm(__StartupFormTemplate):
-    def __init__(self, **properties):
-        # Set Form properties and Data Bindings.
-        self.init_components(**properties)
-        self.tabulator_1.data = [
-            {
-                "columnA": "columnA",
-                "columnB": "columnB",
-                "columnC": "columnC",
-                "columnD": "columnD",
-            },
-            {
-                "columnA": "columnA",
-                "columnB": "columnB",
-                "columnC": "columnC",
-                "columnD": "columnD",
-            },
-        ]
