--- client_code/Tabulator/__StartupForm/form_template.yaml
@@ -1,25 +1,0 @@
-is_package: true
-container:
-  type: HtmlTemplate
-  properties: {html: '@theme:standard-page.html'}
-components:
-- type: FlowPanel
-  properties: {role: null, spacing: medium, align: center, tooltip: '', border: '',
-    foreground: '', visible: true, spacing_above: small, spacing_below: small, background: ''}
-  name: flow_panel_1
-  layout_properties: {slot: default}
-  components:
-  - type: Label
-    properties: {role: null, align: center, tooltip: '', border: '', foreground: '',
-      visible: true, text: This is a dependency, font_size: null, font: '', spacing_above: large,
-      icon_align: left, spacing_below: large, italic: true, background: '', bold: false,
-      underline: false, icon: ''}
-    name: __
-    layout_properties: {slot: default}
-- type: form:Tabulator
-  properties: {pagination: true, index: id, height: '', header_visible: true, pagination_size: 10,
-    header_align: middle, row_selectable: checkbox, pagination_size_selector: 6 10 20,
-    visible: true, layout: fitColumns, auto_columns: true, spacing_above: small, spacing_below: small,
-    movable_columns: false, resizable_columns: true, pagination_button_count: 5}
-  name: tabulator_1
-  layout_properties: {slot: default}
