+++ client_code/Tabulator/__init__.py
@@ -1,440 +1,219 @@
 # SPDX-License-Identifier: MIT
 # Copyright (c) 2022 Stu Cork
 
-"""
-    Tabulator for Anvil
-    an anvil wrapper for tabulator: https://github.com/olifolkerd/tabulator
-    Copyright 2020 Stu Cork
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-        http://www.apache.org/licenses/LICENSE-2.0
-
-    Source code published at https://github.com/s-cork/Tabulator
-"""
-
 from anvil import HtmlTemplate as _HtmlTemplate
 from anvil.js import get_dom_node as _get_dom_node
-from anvil.js.window import Tabulator as _Tabulator
 
+from . import _datetime_overrides
 from ._anvil_designer import TabulatorTemplate
-from ._CleanCols import _clean_cols, _clean_editor, _clean_formatter, _clean_sorter
-from ._Helpers import maintain_scroll_position
+from ._custom_modules import custom_modules
+from ._defaults import (
+    _default_modules,
+    _default_options,
+    _default_props,
+    _default_theme,
+    _event_call_signatures,
+    _methods,
+)
+from ._helpers import (
+    _camelKeys,
+    _ignore_resize_observer_error,
+    _inject_theme,
+    _merge,
+    _options_property,
+    _spacing_property,
+    _to_module,
+    _toCamel,
+)
+from ._js_tabulator import Tabulator as _Tabulator
 
-_default_props = {
-    "auto_columns": None,
-    "header_align": "middle",
-    "header_visible": True,
-    "height": "",
-    "index": "id",
-    "layout": "fitColumns",
-    "movable_columns": False,
-    "pagination": True,
-    "pagination_button_count": 5,
-    "pagination_size": 10,
-    "pagination_size_selector": "6 10 20",
-    "resizable_columns": True,
-    "row_selectable": "checkbox",
-    "spacing_above": "small",
-    "spacing_below": "small",
-    "border": "",
-    "visible": True,
+row_selection_column = {
+    "formatter": "rowSelection",
+    "titleFormatter": "rowSelection",
+    "width": 40,
+    "hozAlign": "center",
+    "headerHozAlign": "center",
+    "headerSort": False,
+    "cellClick": lambda e, cell: cell.getRow().toggleSelect(),
 }
 
 
 class Tabulator(TabulatorTemplate):
+    theme = _default_theme
+    modules = _default_modules
+    default_options = _default_options
+    _registered = False
+
+    def __new__(cls, **properties):
+        cls._setup()
+        return Tabulator5Template.__new__(cls, **properties)
+
     def __init__(self, **properties):
-        # allow Tabulator to be set in code with default values
-        self._from_cache = False
-        self._table_init = False
-
-        el = self._el = _get_dom_node(self)
-        properties = _default_props | properties
-
-        self.init_components(**properties)
-
-        selectable = (
-            "highlight" if self._row_selectable == "checkbox" else self._row_selectable
+        self._t = None
+        self._el = el = _get_dom_node(self)
+        self._queued = []
+        self._handlers = {}
+        self._options = _merge(
+            _default_options, properties, row_formatter=self._row_formatter
         )
 
-        self._table = _Tabulator(
-            el,
-            {
-                "autoColumns": self._auto_columns,
-                "headerAlign": self._header_align,
-                "headerVisible": self._header_visible,
-                "height": self._height,
-                "index": self._index,
-                "layout": self._layout,
-                "moveableColumns": self._resizable_columns,
-                "pagination": self._pagination,
-                "paginationButtonCount": self._pagination_button_count,
-                "paginationSize": self._pagination_size,
-                "paginationSizeSelector": self._pagination_size_selector,
-                "resizableColumns": self._resizable_columns,
-                "selectable": selectable,
-                "rowSelected": self.row_selected,
-                "rowClick": self.row_click,
-                "cellEdited": self.cell_edited,
-                "rowSelectionChanged": self.row_selection_change,
-                "cellClick": self.cell_click,
-                "pageLoaded": self.page_loaded,
-                "rowFormatter": self.row_formatter,
-            },
-        )
+        # public
+        self.options = {}
 
-        self.columns = properties.get("columns", [])
-        self._table_init = True
+        el.replaceChildren()
+        self.init_components(**_merge(_default_props, properties))
 
-    # helpers
-    _clean_cols = _clean_cols
-    _clean_editor = _clean_editor
-    _clean_formatter = _clean_formatter
-    _clean_sorter = _clean_sorter
+    @classmethod
+    def _setup(cls):
+        _inject_theme(cls.theme)
+        if cls._registered:
+            return
+        cls.register_module(cls.modules)
+        cls.register_module(custom_modules)
+        _datetime_overrides.init_overrides()
+        _ignore_resize_observer_error()
+        for key, val in cls.default_options.items():
+            _Tabulator.defaultOptions[key] = val
+        cls._registered = True
 
-    # Methods
+    @staticmethod
+    def register_module(modules):
+        if type(modules) not in (set, tuple, list):
+            modules = [modules]
+        modules = [_to_module(m) for m in modules]
+        _Tabulator.registerModule(modules)
+
+    # because row_formatter is not a tabulator event but it is an anvil tabulator event
+    def _row_formatter(self, row):
+        self.raise_event("row_formatter", row=row)
+
+    def _initialize(self):
+        options = _camelKeys(self._options)
+        options["columns"] = [_camelKeys(defn) for defn in options["columns"]]
+        options["columnDefaults"] = _camelKeys(options["columnDefaults"])
+        options.update(_camelKeys(self.options))
+        data = options.pop("data")
+
+        t = _Tabulator(self._el, options)
+        t.anvil_form = self
+        for attr, args, kws in self._queued:
+            getattr(t, attr)(*args, **kws)
+        self._queued.clear()
+        # use setData - initiating data with anything other than list[dict] breaks tabulator
+
+        def built():
+            t.setData(data)
+            self._t = t
+
+        t.on("tableBuilt", built)
+
+    def _show(self, **event_args):
+        if self._t is None:
+            self._initialize()
+
+    def __getattr__(self, attr):
+        attr = _toCamel(attr)
+        if self._t is None:
+            return lambda *args, **kws: self._queued.append([attr, args, kws])
+        return getattr(self._t, attr)
+
+    def add_event_handler(self, event, handler):
+        super().add_event_handler(event, handler)
+        call_sig = _event_call_signatures.get(event)
+        if call_sig is None:
+            return
+
+        def raiser(*args):
+            kws = dict(zip(call_sig, args))
+            return self.raise_event(event, **kws)
+
+        self._handlers[(event, handler)] = raiser
+        self.on(_toCamel(event), raiser)
+
+    set_event_handler = add_event_handler
+
+    def remove_event_handler(self, event, handler):
+        super().remove_event_handler(event, handler)
+        self.off(_toCamel(event), self._handlers.pop((event, handler), None))
+
+    data = _options_property("data", "getData", "setData")
+    columns = _options_property("columns", None, "setColumns")
+    column_defaults = _options_property("columnDefaults")
+    auto_columns = _options_property("autoColumns")
+    header_visible = _options_property("headerVisible")
+    index = _options_property("index")
+    layout = _options_property("layout")
+    pagination = _options_property("pagination")
+    pagination_size = _options_property("pagination_size", "getPageSize", "setPageSize")
+
+    border = _HtmlTemplate.border
+    visible = _HtmlTemplate.visible
+    role = _HtmlTemplate.role
+    spacing_above = _spacing_property("above")
+    spacing_below = _spacing_property("below")
+
+    #### for the autocomplete
+    def on(self, tabulator_event, handler):
+        """Add an event handler to any tablulator event"""
+
     def add_row(self, row, top=True, pos=None):
         """add a row - ensure the row has an index"""
-        index = row.get(self._index)
-        if index is None:
-            raise KeyError(f"you should provide an index '{self._index}' for this row")
-        if self.get_row(index):
-            raise KeyError(f"The index '{self._index}' should be unique")
-
-        self._table.addRow(row, top, pos)
 
     def delete_row(self, index):
         """delete a row - the index of the row must be provided"""
-        self._table.deleteRow(index)
 
     def update_row(self, index, row):
         """update a row - the index of the row must be provided"""
-        try:
-            self._table.updateRow(index, row)
-        except Exception as e:
-            if "NotFoundError" in repr(e):
-                pass
-            else:
-                raise e
 
     def get_row(self, index):
         """get the row - the index of the row must be provided"""
-        row = self._table.getRow(index)
-        return dict(row.getData()) if row else None
 
     def select_row(self, index_or_indexes):
-        self._table.selectRow(index_or_indexes)
+        """pass the index to select or an array of indexes"""
 
     def deselect_row(self, index_or_indexes=None):
         """deselect a row (single index), rows (list of indexes) or all rows (no argument)"""
-        if index_or_indexes is None:
-            self._table.deselectRow()  # call it with an empty argument
-        else:
-            self._table.deselectRow(index_or_indexes)
 
-    def get_selected(self):
-        """returns a list of selected rows"""
-        return [dict(row) for row in self._table.getSelectedData()]
+    def get_selected_data(self):
+        """returns a list of selected data"""
 
     def add_data(self, data, top=False, pos=None):
         """add data - use the keyword arguments to determine where in the table it gets added"""
-        self._table.addData(data, top, pos)
-
-    def update_or_add_data(self, data):
-        """checks each row and updates data if the row exists, otherwise creates a new row"""
-        self._table.updateOrAddData(data)
-
-    @maintain_scroll_position
-    def replace_data(self, data):
-        # useful to keep the datatable in the same place
-        """replace all data in the table"""
-        self._table.replaceData(data)
-
-    def set_filter(self, field, type=None, value=None):
-        """for multiple filters pass a list of dicts with keys 'field', 'type', 'value'"""
-        if callable(field):
-            filter_func = field
-
-            def field(data, params):
-                return filter_func(dict(data), **params)
-
-        self._table.setFilter(field, type, value)
-
-    def add_filter(self, field, type=None, value=None):
-        if callable(field):
-            filter_func = field
-
-            def field(data, params):
-                return filter_func(dict(data), **params)
-
-        self._table.addFilter(field, type, value)
-
-    def remove_filter(self, field=None, type=None, value=None):
-        self._table.removeFilter(field, type, value)
-
-    def get_filters(self):
-        return self._table.getFilters()
-
-    def clear_filter(self, *args):
-        """include an arg of True to clear header filters as well"""
-        self._table.clearFilter(*args)
-
-    def set_sort(self, column, ascending=True):
-        """first argument can also be a list of sorters [{'column':'name', 'ascending':True}]"""
-        if isinstance(column, list):
-            sorters = column
-            for sorter in sorters:
-                sorter["dir"] = "asc" if sorter.pop("ascending") else "desc"
-            self._table.setSort(sorters)
-        elif isinstance(column, str):
-            self._table.setSort(column, "asc" if ascending else "desc")
-        else:
-            raise TypeError(
-                "expected first argument to be a list of sorters or the column field name"
-            )
-
-    def clear_sort(self):
-        self._table.clearSort()
-        # reset the table data
-        self._table.setData(self._table.getData())
-
-    def set_group_by(self, field):
-        self._table.setGroupBy(field)
-
-    # Events
-    def row_selected(self, row):
-        return self.raise_event("row_selected", row=dict(row.getData()), _row=row)
-
-    def row_click(self, e, row):
-        return self.raise_event("row_click", row=dict(row.getData()), _row=row)
-
-    def row_selection_change(self, e, rows):
-        return self.raise_event(
-            "row_selection_change",
-            rows=[dict(row.getData()) for row in rows],
-            _rows=rows,
-        )
-
-    def cell_click(self, e, cell):
-        return self.raise_event(
-            "cell_click", field=cell.getField(), row=dict(cell.getData(), cell=cell)
-        )
-
-    def page_loaded(self, pageno):
-        return self.raise_event("page_loaded", pageno=pageno)
-
-    def cell_edited(self, cell):
-        return self.raise_event(
-            "cell_edited", field=cell.getField(), row=dict(cell.getData(), cell=cell)
-        )
-
-    def row_formatter(self, row):
-        return self.raise_event("row_formatter", row=row)
-
-    # Lang options
-    def set_locale(self, lang):
-        """set the locale of the table - you must have included this lang option at runtime"""
-        return self._table.setLocale(lang)
-
-    def get_locale(self):
-        """set the current locale of the table"""
-        return self._table.getLocale()
-
-    def get_lang(self, lang=None):
-        if lang is None:
-            return self._table.getLang()
-        return self._table.getLang(lang)
-
-    @property
-    def langs(self):
-        """lang_options is a dictionary of language options eg. {'fr': {...options}, 'de': {...options}}"""
-        raise AttributeError(
-            "cannot read the lang property use get_lang(lang) to get a specific language option"
-        )
-
-    @langs.setter
-    def langs(self, lang_options):
-        if not self._table_init:
-            raise RuntimeError(
-                "lang property must be set at runtime not at initialization"
-            )
-        for locale, lang in lang_options.items():
-            self._table.modules.localize.installLang(locale, lang)
-
-    @maintain_scroll_position
-    def redraw(self, full_render=False, **event_args):
-        """This method is called when the HTML panel is shown on the screen"""
-        self._table.redraw(full_render)
-
-    def form_show(self, **event_args):
-        # redraw on show
-        self.redraw(full_render=not self._from_cache)
-        self._from_cache = True
 
     def get_data(self, active="all"):
         """
         Returns the table data based on a Tabulator range row lookup value.
         :active: Range row lookup. Valid values are: "visible", "active", "selected", "all"
         """
-        return [dict(row) for row in self._table.getData(active)]
 
-    # properties
-    data = property(get_data)
+    def update_or_add_data(self, data):
+        """checks each row and updates data if the row exists, otherwise creates a new row"""
 
-    @data.setter
-    def data(self, value):
-        self._table.setData(value)
+    def replace_data(self, data):
+        """replace all data in the table"""
 
-    @property
-    def height(self):
-        return self._height
+    def set_filter(self, field, type=None, value=None):
+        """for multiple filters pass a list of dicts with keys 'field', 'type', 'value'
+        Can also pass in a function and dictionary of keyword arguments to pass to that function"""
 
-    @height.setter
-    def height(self, value):
-        self._height = value
-        if value.isdigit():
-            value = value + "px"
-        if self.parent:
-            self._table.setHeight(value)
+    def add_filter(self, field, type=None, value=None):
+        """Add a filter to"""
 
-    @property
-    def columns(self):
-        return self._columns or []
+    def remove_filter(self, field=None, type=None, value=None):
+        """remove a filter"""
 
-    @columns.setter
-    def columns(self, value):
-        self._columns = value
-        cols = self._clean_cols(value)
-        self._table.setColumns(cols)
+    def get_filters(self):
+        """get a list of the current filters"""
 
-    @property
-    def index(self):
-        return self._index
+    def clear_filter(self, clear_header=False):
+        """include an arg of True to clear header filters as well"""
 
-    @index.setter
-    def index(self, value):
-        self._index = value
+    def set_sort(self, column, dir):
+        """first argument can also be a list of sorters [{'column': field, 'dir':'asc' | 'desc'}, ...]"""
 
-    @property
-    def auto_columns(self):
-        return self._auto_columns
+    def clear_sort(self):
+        """clear the sorters"""
 
-    @auto_columns.setter
-    def auto_columns(self, value):
-        self._auto_columns = bool(value)
 
-    @property
-    def resizable_columns(self):
-        return self._resizable_columns
-
-    @resizable_columns.setter
-    def resizable_columns(self, value):
-        self._resizable_columns = bool(value)
-
-    @property
-    def movable_columns(self):
-        return self._movable_columns
-
-    @movable_columns.setter
-    def movable_columns(self, value):
-        self._movable_columns = bool(value)
-
-    @property
-    def header_align(self):
-        return self._header_align
-
-    @header_align.setter
-    def header_align(self, value):
-        value = value.strip().lower()
-        if value not in ("top", "middle", "bottom"):
-            value = "middle"
-        self._header_align = value
-
-    @property
-    def header_visible(self):
-        return self._header_visible
-
-    @header_visible.setter
-    def header_visible(self, value):
-        self._header_visible = bool(value)
-
-    @property
-    def pagination(self):
-        return self._pagination
-
-    @pagination.setter
-    def pagination(self, value):
-        self._pagination = "local" if value else value
-
-    @property
-    def pagination_size(self):
-        return self._pagination_size
-
-    @pagination_size.setter
-    def pagination_size(self, value):
-        self._pagination_size = int(value)
-
-    @property
-    def pagination_button_count(self):
-        return self._pagination_button_count
-
-    @pagination_button_count.setter
-    def pagination_button_count(self, value):
-        self._pagination_button_count = int(value)
-
-    @property
-    def row_selectable(self):
-        return self._row_selectable
-
-    @row_selectable.setter
-    def row_selectable(self, value):
-        """true, false, checkbox, highlight"""
-        self._row_selectable = value
-
-    @property
-    def pagination_size_selector(self):
-        return self._pagination_size_selector
-
-    @pagination_size_selector.setter
-    def pagination_size_selector(self, value):
-        value = value.strip()
-        if value == "None" or not value:
-            self._pagination_size_selector = None
-        else:
-            value = value.replace(",", " ")
-            self._pagination_size_selector = [int(i) for i in value.split()]
-
-    @property
-    def layout(self):
-        return self._layout
-
-    @layout.setter
-    def layout(self, value):
-        layouts = ("fitData", "fitDataFill", "fitDataStretch", "fitColumns")
-        if value not in layouts:
-            value = "fitColumns"
-        self._layout = value
-
-    @property
-    def spacing_above(self):
-        return self._spacing_above
-
-    @spacing_above.setter
-    def spacing_above(self, value):
-        self._spacing_above = value
-        self._el.classList.add("anvil-spacing-above-" + self._spacing_above)
-
-    @property
-    def spacing_below(self):
-        return self._spacing_below
-
-    @spacing_below.setter
-    def spacing_below(self, value):
-        self._spacing_below = value
-        self._el.classList.add("anvil-spacing-below-" + self._spacing_below)
-
-    border = _HtmlTemplate.border
-    visible = _HtmlTemplate.visible
-    role = _HtmlTemplate.role
+for method in _methods:
+    delattr(Tabulator5, method)
