+++ client_code/Tabulator/form_template.yaml
@@ -1,57 +1,55 @@
 properties:
+- {name: auto_columns, type: boolean, default_value: null, group: columns, important: false}
+- {name: header_visible, type: boolean, default_value: true, group: header, important: true}
 - {name: height, type: string, default_value: '', default_binding_prop: true, group: appearance,
   important: false}
-- {name: data, type: object, default_value: null, important: true, description: typically a list of dicts}
-- {name: layout, type: string, default_value: fitColumns, group: columns, important: false}
-- {name: columns, type: object, default_value: null, important: true, description: column definitions should be a list of dicts. Each dict represents a column with options. See Tabulator.info for full documentation on available options}
-- {name: pagination_size_selector, type: string, default_value: 6 10 20, group: pagination,
-  important: false}
 - {name: index, type: string, default_value: id, important: true, description: This should be a unique identifier for a row.}
-- {name: auto_columns, type: boolean, default_value: null, group: columns, important: false}
-- {name: resizable_columns, type: boolean, default_value: true, group: columns, important: false}
-- {name: movable_columns, type: boolean, default_value: false, group: columns, important: false}
-- {name: header_align, type: string, default_value: middle, group: header, important: false}
-- {name: header_visible, type: boolean, default_value: true, group: header, important: true}
-- {name: pagination, type: boolean, default_value: true, group: pagination, important: true}
-- {name: pagination_size, type: number, default_value: 10, group: pagination, important: false}
-- {name: pagination_button_count, type: number, default_value: 5, group: pagination,
-  important: false}
-- {name: row_selectable, type: string, default_value: checkbox, group: appearance,
-  important: false}
+- {name: pagination, type: boolean, default_value: true, default_binding_prop: false}
+- {name: pagination_size, type: number, default_value: 5, group: pagination, important: false}
+- {name: border, type: string, default_value: '', group: appearance, important: false}
 - {name: spacing_above, type: string, default_value: small, group: layout, important: false}
 - {name: spacing_below, type: string, default_value: small, group: layout, important: false}
 - {name: visible, type: boolean, default_value: true, group: appearance, important: true}
-- {name: border, type: string, default_value: '', group: appearance, important: false}
-- {name: langs, type: object, default_value: null, description: set lang options,
-  important: false}
+- {name: layout, type: string, default_value: fitColumns, group: columns, important: false,
+  description: 'valid values include fitColumns, fitData, fitDataTable, fitDataStretch
+    and fitDataFill. See tabulator docs for examples'}
 - {name: role, type: string, default_value: '', group: appearance, important: false}
+- {name: data, type: object, default_value: null, important: true, description: typically a list of dicts}
+- {name: columns, type: object, default_value: null, important: true, description: column definitions should be a list of dicts. Each dict represents a column with options. See Tabulator.info for full documentation on available options}
+- {name: column_defaults, type: object, default_value: null, description: set to a dict at runtime}
 is_package: true
 events:
 - name: row_click
   default_event: true
   description: when a row is clicked
   parameters:
-  - {name: row, description: ''}
+  - {name: row, description: the tabulator row object}
+- name: row_tap
+  default_event: true
+  description: when a row is tapped on a mobile device
+  parameters:
+  - {name: row, description: the tabulator row object}
+- name: cell_click
+  description: when a cell is clicked
+  parameters:
+  - {name: cell, description: the tabulator cell object}
+- name: cell_tap
+  description: when a cell is tapped on a mobile device
+  parameters:
+  - {name: cell, description: the tabulator cell object}
 - name: row_selected
   description: when a row is selected
   parameters:
-  - {name: row, description: ''}
+  - {name: row, description: the tabulator row object}
 - name: row_selection_change
   parameters:
   - {name: rows}
+  - {name: data}
   description: when the row selection changes
-- name: cell_click
-  description: when a cell is clicked - event_args include field and row
+- name: row_deselected
+  description: when a row is deselected
   parameters:
-  - {name: field}
-  - {name: row}
-  - {name: cell, description: the tabulator cell object}
-- name: cell_edited
-  description: when a cell is edited - event_args include field and row
-  parameters:
-  - {name: field}
-  - {name: row}
-  - {name: cell, description: the tabulator cell object}
+  - {name: row, description: the tabulator row object}
 - name: page_loaded
   description: when a page is loaded
   parameters:
@@ -67,11 +65,12 @@ components: []
 container:
   type: HtmlTemplate
   properties: {tooltip: '', background: '', foreground: '', border: '', visible: true,
-    role: null, html: "<div class='placeholder-table'></div>\n\n<script>\n  \n$('.placeholder-table').each(function()\
-      \ {\n    table = new Tabulator(this, {\n    layout:\"fitColumns\",\n    pagination:\"\
-      local\",\n    paginationSize:6,\n    paginationSizeSelector:[3, 6, 10],\n  \
-      \  autoColumns: true,\n  \tdata: [{columnA:'columnA', columnB:'columnB', columnC:'columnC',\
-      \ columnD:'columnD'},\n          {columnA:'columnA', columnB:'columnB', columnC:'columnC',\
-      \ columnD:'columnD'}],\n\t});\n\n});\n</script>\n\n<style>\n  .tabulator-row\
-      \ .tabulator-cell {\n    font-style: italic;\n  }\n</style>\n\n"}
-  event_bindings: {show: form_show}
+    role: null, html: "<div class='placeholder-table'></div>\n\n<script type=\"module\"\
+      >\nimport {TabulatorFull as Tabulator} from \"https://cdn.skypack.dev/tabulator-tables@5.1.2\"\
+      ;\n\n$('.placeholder-table').each(function() {\n    new Tabulator(this, {\n\
+      \          autoColumns: true,\n          layout:\"fitColumns\",\n          pagination:\
+      \ true,\n          paginationSize:2,\n          data: new Array(6).fill({columnA:'columnA',\
+      \ columnB:'columnB', columnC:'columnC', columnD:'columnD'})\n    });\n});\n\
+      </script>\n\n<style>\n.tabulator-row .tabulator-cell {\n    font-style: italic;\n\
+      }\n</style>\n\n"}
+  event_bindings: {show: _show}
