--- client_code/Tabulator5/form_template.yaml
@@ -1,74 +1,0 @@
-properties:
-- {name: auto_columns, type: boolean, default_value: null, group: columns, important: false}
-- {name: header_visible, type: boolean, default_value: true, group: header, important: true}
-- {name: height, type: string, default_value: '', default_binding_prop: true, group: appearance,
-  important: false}
-- {name: index, type: string, default_value: id, important: true, description: This should be a unique identifier for a row.}
-- {name: pagination, type: boolean, default_value: true, default_binding_prop: false}
-- {name: pagination_size, type: number, default_value: 5, group: pagination, important: false}
-- {name: border, type: string, default_value: '', group: appearance, important: false}
-- {name: spacing_above, type: string, default_value: small, group: layout, important: false}
-- {name: spacing_below, type: string, default_value: small, group: layout, important: false}
-- {name: visible, type: boolean, default_value: true, group: appearance, important: true}
-- {name: layout, type: string, default_value: fitColumns, group: columns, important: false, description: "valid values include fitColumns, fitData, fitDataTable, fitDataStretch and fitDataFill. See tabulator docs for examples"}
-- {name: role, type: string, default_value: '', group: appearance, important: false}
-- {name: data, type: object, default_value: null, important: true, description: typically a list of dicts}
-- {name: columns, type: object, default_value: null, important: true, description: column definitions should be a list of dicts. Each dict represents a column with options. See Tabulator.info for full documentation on available options}
-- {name: column_defaults, type: object, default_value: null, description: set to a dict at runtime}
-is_package: true
-events:
-- name: row_click
-  default_event: true
-  description: when a row is clicked
-  parameters:
-  - {name: row, description: the tabulator row object}
-- name: row_tap
-  default_event: true
-  description: when a row is tapped on a mobile device
-  parameters:
-  - {name: row, description: the tabulator row object}
-- name: cell_click
-  description: when a cell is clicked
-  parameters:
-  - {name: cell, description: the tabulator cell object}
-- name: cell_tap
-  description: when a cell is tapped on a mobile device
-  parameters:
-  - {name: cell, description: the tabulator cell object}
-- name: row_selected
-  description: when a row is selected
-  parameters:
-  - {name: row, description: the tabulator row object}
-- name: row_selection_change
-  parameters:
-  - {name: rows}
-  - {name: data}
-  description: when the row selection changes
-- name: row_deselected
-  description: when a row is deselected
-  parameters:
-  - {name: row, description: the tabulator row object}
-- name: page_loaded
-  description: when a page is loaded
-  parameters:
-  - {name: pageno}
-- name: row_formatter
-  description: when the row is rendered - manipulate the tabulator row object
-  parameters:
-  - {name: row, description: the tabulator row object}
-- {name: show, description: when the tabulator is shown}
-- {name: hide, description: when the tabulator is hidden}
-custom_component: true
-components: []
-container:
-  type: HtmlTemplate
-  properties: {tooltip: '', background: '', foreground: '', border: '', visible: true,
-    role: null, html: "<div class='placeholder-table'></div>\n\n<script type=\"module\"\
-      >\nimport {TabulatorFull as Tabulator} from \"https://cdn.skypack.dev/tabulator-tables@5.1.2\"\
-      ;\n\n$('.placeholder-table').each(function() {\n    new Tabulator(this, {\n\
-      \          autoColumns: true,\n          layout:\"fitColumns\",\n          pagination:\
-      \ true,\n          paginationSize:2,\n          data: new Array(6).fill({columnA:'columnA',\
-      \ columnB:'columnB', columnC:'columnC', columnD:'columnD'})\n    });\n});\n\
-      </script>\n\n<style>\n.tabulator-row .tabulator-cell {\n    font-style: italic;\n\
-      }\n</style>\n\n"}
-  event_bindings: {show: _show}
